add Unit Tests For Services

+++ src/main/java/com/example/demo/services/LineItemService.java
@@ -19,15 +19,13 @@ public class LineItemService {
 @Autowired
     public LineItemService(LineItemRepository lineItemRepository, ProductRepository productRepository, CartRepository cartRepository) {
         this.lineItemRepository = lineItemRepository;
-    this.productRepository = productRepository;
-    this.cartRepository = cartRepository;
+        this.productRepository = productRepository;
+        this.cartRepository = cartRepository;
 }
     public LineItemResponse getLineItem(Long lineItemId) {
         boolean exist = lineItemRepository.existsById(lineItemId);
         if (!exist)
-            throw new IllegalStateException(
-                    "Line Item with id " + lineItemId + " doesn't exist"
-            );
+            throw new IllegalStateException("Line Item with id " + lineItemId + " doesn't exist");
         LineItemResponse response = new LineItemResponse();
         LineItem lineItem = lineItemRepository.findById(lineItemId).get();
         response.setQuantity(lineItem.getQuantity());
@@ -41,15 +39,13 @@ public class LineItemService {
         lineItemRepository.save(lineItem);
     }
     public void deleteLineItem(Long lineItemId) {
-        LineItem lineItem = lineItemRepository.findById(lineItemId)
-                .orElseThrow(() -> new IllegalStateException(
-                        "Line item with id " + lineItemId + " doesn't exist"));
+        boolean exist = lineItemRepository.existsById(lineItemId);
+        if (!exist)
+            throw new IllegalStateException("Line Item with id " + lineItemId + " doesn't exist");
         lineItemRepository.deleteById(lineItemId);
     }
-
     public void updateProduct(Long lineItemId, LineItemRequest request) {
-        LineItem lineItem = lineItemRepository.findById(lineItemId)
-                .orElseThrow(() -> new IllegalStateException(
+        LineItem lineItem = lineItemRepository.findById(lineItemId).orElseThrow(() -> new IllegalStateException(
                         "Line item with id " + lineItemId + " doesn't exist"));
         if (request.getQuantity() > 0)
             lineItem.setQuantity(request.getQuantity());
@@ -58,18 +54,12 @@ public class LineItemService {
     public void putProductInLineItem(ProductLineItemRequest productLineItemRequest) {
         if (!productRepository.existsById(productLineItemRequest.getProductId()))
             throw new IllegalStateException("Product with id " + productLineItemRequest.getProductId() + " doesn't exist");
-
         LineItem lineItem = new LineItem();
         Product product = productRepository.findById(productLineItemRequest.getProductId()).get();
         lineItem.setProduct(product);
         lineItem.setQuantity(productLineItemRequest.getQuantity());
-
-        putLineItemInCart(productLineItemRequest, lineItem);
-        lineItemRepository.save(lineItem);
-    }
-
-    private void putLineItemInCart(ProductLineItemRequest productLineItemRequest, LineItem lineItem) {
         Cart cart=  cartRepository.findById(productLineItemRequest.getCartId()).get();
         lineItem.setCart(cart);
+        lineItemRepository.save(lineItem);
     }
 }
